+++ thesettlements-api/app/auth.py
@@ -4,9 +4,9 @@ from datetime import datetime, timedelta
 from fastapi.security import OAuth2PasswordBearer
 from jose import jwt
 
-from app.core.config import settings
-from app.core.security import verify_password
-from app.schemas.user import User, UserInDB
+from app.config import settings
+from app.utils.security import verify_password
+from app.users.schemas import UserRead, UserInDB
 
 
 JWTPayloadMapping = MutableMapping[str, datetime | bool | str | list[str] | list[int]]
@@ -14,7 +14,7 @@ JWTPayloadMapping = MutableMapping[str, datetime | bool | str | list[str] | list
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl=f"{settings.API_V_LATEST_STR}/auth/login")
 
 
-async def authenticate(username: str, password: str, collection) -> User | None:
+async def authenticate(username: str, password: str, collection) -> UserRead | None:
     user = await collection.find_one({"username": username})
     if not user:
         return None
@@ -22,12 +22,14 @@ async def authenticate(username: str, password: str, collection) -> User | None:
         return None
     return user
 
+
 def create_access_token(sub: str) -> str:
     return _create_token(
         token_type="access_token",
         lifetime=timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES),
         sub=sub,
     )
+
 
 def _create_token(token_type: str, lifetime: timedelta, sub: str) -> str:
     payload = {}