+++ thesettlements-api/app/settlements/schemas.py
@@ -1,10 +1,17 @@
+from uuid import UUID
+
 from pydantic import BaseModel, Field
-from bson import ObjectId
 
-from app.utils.mongo_utils import PyObjectId
-from app.schemas.structure import Structure
+from app.structures.schemas import Structure
 
-class Settlement(BaseModel):
+
+class SettlementBase(BaseModel):
+    name: str
+    x: int
+    y: int
+    population: int | None = 100
+    structures: list[Structure] | None = Field(default_factory=list)
+    owner: UUID | None = None
     id: PyObjectId = Field(default_factory=PyObjectId, alias="_id")
     name: str
     x: int
@@ -12,11 +19,11 @@ class Settlement(BaseModel):
     population: int = 100
     structures: list[Structure] = []
     owner: PyObjectId | None = Field(default_factory=PyObjectId)
-       
+
     class Config:
         allow_population_by_field_name = True
         arbitrary_types_allowed = True
-        json_encoders = { ObjectId: str }
+        json_encoders = {ObjectId: str}
         schema_extra = {
             "example": {
                 "name": "Roma",
@@ -27,17 +34,8 @@ class Settlement(BaseModel):
             }
         }
 
-class SettlementUpdate(BaseModel):
-    name: str | None
-    population: int | None
-    owner: PyObjectId | None
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "name": "Roma",
-                "population": 1000,
-                "owner": "1",
-            }
-        }
-
+class SettlementUpdate(SettlementBase):
+    name: str | None = None
+    population: int | None = None
+    structures: list[Structure] | None = None