+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # 0.1 - project refactor
 
+### [0.1.2] - 2023-09-03
+- `schemas` refactor and small cleaning
+
 ### [0.1.1] - 2023-09-02
-- session and settings refactor
+- `session` and `settings` refactor
 
 ### [0.1.0] - 2023-08-29
 - general cleaning